Reorganize HomeBox integration into connector/importer/shared layout
Pure code-organization refactor preparing the HomeBox integration for
the Connect/Import split (Issue #354 and #355). No behavior changes.

HbConverter is split: pure data-shaping helpers stay in
shared/hb_converter.py; the six DB-write classmethods move to a new
HbImporter class at importer/hb_importer.py and are called from
importer/hb_sync.py. The connector/ subpackage is scaffolded empty
for later phases.
