+++ src/hi/services/homebox/shared/hb_client.py
